requirements.txt
enum34==1.1.10
fire==0.3.1
future==0.18.2
numpy==1.16.6
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
Pillow==6.2.2
six==1.14.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
